+++ src/main/kotlin/org/yelog/ideavim/flash/UserConfig.kt
@@ -32,6 +32,7 @@ class UserConfig : PersistentStateComponent<UserConfig.DataBean> {
         var scrolloff = DEFAULT_SCROLL_OFF
         var searchAcrossSplits = DEFAULT_SEARCH_ACROSS_SPLITS
         var vimModeTimeoutMillis = DEFAULT_VIM_MODE_TIMEOUT
+        var enterConfirmsNearestMatch = DEFAULT_ENTER_CONFIRMS_NEAREST_MATCH
     }
 
     companion object {
@@ -49,6 +50,7 @@ class UserConfig : PersistentStateComponent<UserConfig.DataBean> {
         const val DEFAULT_SCROLL_OFF = 4
         const val DEFAULT_SEARCH_ACROSS_SPLITS = true
         const val DEFAULT_VIM_MODE_TIMEOUT = -1
+        const val DEFAULT_ENTER_CONFIRMS_NEAREST_MATCH = true
 
         // get instance
         private fun getInstance(): UserConfig {